Moved analyse and build regression into one script
