--- examples/data/elasticity_tensor_workflow/defor_1/defor_1.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.556143821003581 -1.9583704e-09 9.889083e-10 -1.9776713e-09 3.521109146866583 2.2204145e-08 9.987171e-10 2.22055362e-08 3.5211077603179106" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.2554122297458 stress="0.019088708368710874 -6.817172905517905e-10 3.4425484722946284e-10 -6.817172905517905e-10 0.013850121843948803 8.051617891374034e-09 3.4425484722946284e-10 8.051617891374034e-09 0.013849933124702056" free_energy=-28.2554122297458 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co      -0.00000000       1.76055458       1.76055389       3.00000000      -0.00000000       0.00000000      -0.00000000
-Co       1.77807191       0.00000001       1.76055388       3.00000000       0.00000000      -0.00000000      -0.00000000
-Co       1.77807191       1.76055457       0.00000001       3.00000000       0.00000000       0.00000000       0.00000000
