--- examples/data/elasticity_tensor_workflow/defor_10/defor_10.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.5211070668977125 -1.9583704e-09 9.871267e-10 -6.259377e-10 3.521109146866583 2.21641415e-08 0.2112664666079519 2.22055362e-08 3.514764051864056" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.207973437175568 stress="-0.00707984389165707 1.1023374702445959e-10 0.03905602107320268 1.1023374702445959e-10 -0.0041070774463651035 8.197294691482791e-09 0.03905602107320268 8.197294691482791e-09 -0.011774917217889128" free_energy=-28.207973437175568 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co       0.10563323       1.76055458       1.75738204       3.00000000      -0.00000000       0.00000000      -0.00000000
-Co       1.86618677       0.00000001       1.75738203       3.00000000       0.00000000      -0.00000000      -0.00000000
-Co       1.76055353       1.76055457       0.00000001       3.00000000       0.00000000       0.00000000       0.00000000
