--- examples/data/elasticity_tensor_workflow/defor_10/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_10/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.5211070668977125 -0.0000000019583704 0.0000000009871267 
-lattice_vector -0.0000000006259377 3.5211091468665829 0.0000000221641415 
-lattice_vector 0.2112664666079519 0.0000000222055362 3.5147640518640562 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom 0.1056332329910071 1.7605545845360591 1.7573820370140989 Co
-    initial_moment         3.000000
-atom 1.8661867667528318 0.0000000101235829 1.7573820264255915 Co
-    initial_moment         3.000000
-atom 1.7605535331358868 1.7605545724541063 0.0000000115756341 Co
-    initial_moment         3.000000
