--- examples/data/elasticity_tensor_workflow/defor_11/defor_11.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066838378 -1.9385922e-09 9.887105e-10 -1.9581864e-09 3.521109147310666 2.21997037e-08 9.888773e-10 0.0704221774118944 3.520403468329604" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.253939799954814 stress="-0.00044341426249494044 -7.132098985216834e-10 3.5531541687162183e-10 -7.132098985216834e-10 -0.0007650324120449628 0.013051105885655154 3.5531541687162183e-10 0.013051105885655154 -0.0012873928152431823" free_energy=-28.253939799954814 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co      -0.00000000       1.79576566       1.76020175       3.00000000      -0.00000000       0.00000000       0.00000000
-Co       1.76055353       0.03521109       1.76020173       3.00000000       0.00000000      -0.00000000       0.00000000
-Co       1.76055353       1.76055457       0.00000001       3.00000000       0.00000000       0.00000000      -0.00000000
