--- examples/data/elasticity_tensor_workflow/defor_11/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_11/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.5211070668383782 -0.0000000019385922 0.0000000009887105 
-lattice_vector -0.0000000019581864 3.5211091473106659 0.0000000221997037 
-lattice_vector 0.0000000009888773 0.0704221774118944 3.5204034683296039 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom -0.0000000004846546 1.7957656623612797 1.7602017452646539 Co
-    initial_moment         3.000000
-atom 1.7605535339136273 0.0352110877366511 1.7602017346591572 Co
-    initial_moment         3.000000
-atom 1.7605535324400954 1.7605545726860368 0.0000000115942071 Co
-    initial_moment         3.000000
