--- examples/data/elasticity_tensor_workflow/defor_12/defor_12.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066838378 -1.8990359e-09 9.871267e-10 -1.9581864e-09 3.5211091481988315 2.21641415e-08 9.888773e-10 0.2112664878246108 3.514764051864056" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.207973553578192 stress="-0.004107392726360672 -6.91375124540424e-10 3.3609736941446193e-10 -6.91375124540424e-10 -0.0070795233180712454 0.03905592307130941 3.3609736941446193e-10 0.03905592307130941 -0.011774909247402472" free_energy=-28.207973553578192 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co      -0.00000000       1.86618782       1.75738204       3.00000000      -0.00000000       0.00000000      -0.00000000
-Co       1.76055353       0.10563324       1.75738203       3.00000000       0.00000000      -0.00000000      -0.00000000
-Co       1.76055353       1.76055457       0.00000001       3.00000000       0.00000000       0.00000000       0.00000000
