--- examples/data/elasticity_tensor_workflow/defor_12/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_12/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.5211070668383782 -0.0000000018990359 0.0000000009871267 
-lattice_vector -0.0000000019581864 3.5211091481988315 0.0000000221641415 
-lattice_vector 0.0000000009888773 0.2112664878246108 3.5147640518640562 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom -0.0000000004846546 1.8661878180117206 1.7573820370140989 Co
-    initial_moment         3.000000
-atom 1.7605535339136273 0.1056332429627875 1.7573820264255915 Co
-    initial_moment         3.000000
-atom 1.7605535324400954 1.7605545731498977 0.0000000115756341 Co
-    initial_moment         3.000000
