--- examples/data/elasticity_tensor_workflow/defor_2/defor_2.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.6252016046983533 -1.9583704e-09 9.889083e-10 -2.0160763e-09 3.521109146866583 2.2204145e-08 1.0181115e-09 2.22055362e-08 3.5211077603179106" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.225175068543336 stress="0.0501392156512595 -5.994924819918262e-10 3.0273298581890437e-10 -5.994924819918262e-10 0.03548427462352261 7.698323317047583e-09 3.0273298581890437e-10 7.698323317047583e-09 0.03548412667416048" free_energy=-28.225175068543336 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co      -0.00000000       1.76055458       1.76055389       3.00000000      -0.00000000       0.00000000      -0.00000000
-Co       1.81260080       0.00000001       1.76055388       3.00000000       0.00000000      -0.00000000      -0.00000000
-Co       1.81260080       1.76055457       0.00000001       3.00000000       0.00000000      -0.00000000       0.00000000
