--- examples/data/elasticity_tensor_workflow/defor_2/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_2/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.6252016046983533 -0.0000000019583704 0.0000000009889083 
-lattice_vector -0.0000000020160763 3.5211091468665829 0.0000000222041450 
-lattice_vector 0.0000000010181115 0.0000000222055362 3.5211077603179106 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom -0.0000000004989824 1.7605545845360591 1.7605538912610279 Co
-    initial_moment         3.000000
-atom 1.8126008028582319 0.0000000101235829 1.7605538806534096 Co
-    initial_moment         3.000000
-atom 1.8126008013411379 1.7605545724541063 0.0000000115965267 Co
-    initial_moment         3.000000
