--- examples/data/elasticity_tensor_workflow/defor_3/defor_3.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066838378 -1.9778571e-09 9.889083e-10 -1.9581864e-09 3.556145921729094 2.2204145e-08 9.888773e-10 2.24264923e-08 3.5211077603179106" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.25541209329755 stress="0.013849807839160995 -6.817176783253627e-10 3.5857991614819695e-10 -6.817176783253627e-10 0.01908901817598288 7.72996112544023e-09 3.5857991614819695e-10 7.72996112544023e-09 0.013849902226128585" free_energy=-28.25541209329755 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000      -0.00000000
-Co      -0.00000000       1.77807297       1.76055389       3.00000000      -0.00000000       0.00000000      -0.00000000
-Co       1.76055353       0.00000001       1.76055388       3.00000000       0.00000000      -0.00000000       0.00000000
-Co       1.76055353       1.77807296       0.00000001       3.00000000       0.00000000       0.00000000       0.00000000
