--- examples/data/elasticity_tensor_workflow/defor_4/defor_4.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066838378 -2.0162657e-09 9.889083e-10 -1.9581864e-09 3.625203746218461 2.2204145e-08 9.888773e-10 2.28619988e-08 3.5211077603179106" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.225174679432268 stress="0.03548396678075803 -5.994933368649792e-10 3.4284572969744557e-10 -5.994933368649792e-10 0.05013949359327621 6.797622738343988e-09 3.4284572969744557e-10 6.797622738343988e-09 0.035484040776782494" free_energy=-28.225174679432268 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co      -0.00000000       1.81260188       1.76055389       3.00000000      -0.00000000       0.00000000      -0.00000000
-Co       1.76055353       0.00000001       1.76055388       3.00000000       0.00000000      -0.00000000       0.00000000
-Co       1.76055353       1.81260187       0.00000001       3.00000000       0.00000000       0.00000000       0.00000000
