--- examples/data/elasticity_tensor_workflow/defor_4/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_4/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.5211070668383782 -0.0000000020162657 0.0000000009889083 
-lattice_vector -0.0000000019581864 3.6252037462184612 0.0000000222041450 
-lattice_vector 0.0000000009888773 0.0000000228619988 3.5211077603179106 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom -0.0000000004846546 1.8126018845402296 1.7605538912610279 Co
-    initial_moment         3.000000
-atom 1.7605535339136273 0.0000000104228665 1.7605538806534096 Co
-    initial_moment         3.000000
-atom 1.7605535324400954 1.8126018721010977 0.0000000115965267 Co
-    initial_moment         3.000000
