--- examples/data/elasticity_tensor_workflow/defor_5/defor_5.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066838378 -1.9583704e-09 9.987484e-10 -1.9581864e-09 3.521109146866583 2.24250872e-08 9.888773e-10 2.22055362e-08 3.5561445213835774" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.255412184254233 stress="0.013849828436182161 -7.10084901164754e-10 3.4425480813938475e-10 -7.10084901164754e-10 0.013850111542396543 7.729958323837845e-09 3.4425480813938475e-10 7.729958323837845e-09 0.019088811657847395" free_energy=-28.255412184254233 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co      -0.00000000       1.76055458       1.77807227       3.00000000      -0.00000000       0.00000000       0.00000000
-Co       1.76055353       0.00000001       1.77807226       3.00000000       0.00000000      -0.00000000       0.00000000
-Co       1.76055353       1.76055457       0.00000001       3.00000000       0.00000000       0.00000000      -0.00000000
