--- examples/data/elasticity_tensor_workflow/defor_6/defor_6.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066838378 -1.9583704e-09 1.0181434e-09 -1.9581864e-09 3.521109146866583 2.28605664e-08 9.888773e-10 2.22055362e-08 3.625202318679231" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.22517493881445 stress="0.03548402404004113 -6.789273494105933e-10 3.0273302951339136e-10 -6.789273494105933e-10 0.035484245985427675 6.797616394418284e-09 3.0273302951339136e-10 6.797616394418284e-09 0.050139308316646286" free_energy=-28.22517493881445 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co      -0.00000000       1.76055458       1.81260117       3.00000000      -0.00000000       0.00000000       0.00000000
-Co       1.76055353       0.00000001       1.81260116       3.00000000       0.00000000      -0.00000000      -0.00000000
-Co       1.76055353       1.76055457       0.00000001       3.00000000       0.00000000      -0.00000000      -0.00000000
