--- examples/data/elasticity_tensor_workflow/defor_6/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_6/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.5211070668383782 -0.0000000019583704 0.0000000010181434 
-lattice_vector -0.0000000019581864 3.5211091468665829 0.0000000228605664 
-lattice_vector 0.0000000009888773 0.0000000222055362 3.6252023186792308 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom -0.0000000004846546 1.7605545845360591 1.8126011707698986 Co
-    initial_moment         3.000000
-atom 1.7605535339136273 0.0000000101235829 1.8126011598486871 Co
-    initial_moment         3.000000
-atom 1.7605535324400954 1.7605545724541063 0.0000000119393550 Co
-    initial_moment         3.000000
