--- examples/data/elasticity_tensor_workflow/defor_7/defor_7.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066799211 -1.9579787e-09 9.889083e-10 0.0704221809791452 3.5204048546009385 2.2204145e-08 1.432988e-09 2.22010947e-08 3.5211077603179106" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.253939802762837 stress="-0.0007653490381638118 0.013051112682434057 6.586093832811039e-10 0.013051112682434057 -0.0012871803635807206 8.235922256240817e-09 6.586093832811039e-10 8.235922256240817e-09 -0.0004433084265134064" free_energy=-28.253939802762837 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co       0.03521109       1.76020244       1.76055389       3.00000000      -0.00000000       0.00000000       0.00000000
-Co       1.76055353       0.00000001       1.76055388       3.00000000       0.00000000      -0.00000000      -0.00000000
-Co       1.79576462       1.76020243       0.00000001       3.00000000       0.00000000       0.00000000      -0.00000000
