--- examples/data/elasticity_tensor_workflow/defor_7/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_7/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.5211070667992108 -0.0000000019579787 0.0000000009889083 
-lattice_vector 0.0704221809791452 3.5204048546009385 0.0000000222041450 
-lattice_vector 0.0000000014329880 0.0000000222010947 3.5211077603179106 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom 0.0352110912060666 1.7602024384010162 1.7605538912610279 Co
-    initial_moment         3.000000
-atom 1.7605535341160989 0.0000000101215580 1.7605538806534096 Co
-    initial_moment         3.000000
-atom 1.7957646238891773 1.7602024263214799 0.0000000115965267 Co
-    initial_moment         3.000000
