--- examples/data/elasticity_tensor_workflow/defor_8/defor_8.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066720876 -1.9548421e-09 9.889083e-10 0.2112665468538086 3.5147654359146907 2.2204145e-08 2.3212095e-09 2.21655302e-08 3.5211077603179106" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.207973531320143 stress="-0.007079844012415896 0.039055962531080506 1.206972237431659e-09 0.039055962531080506 -0.011774694065509492 8.242928361733151e-09 1.206972237431659e-09 8.242928361733151e-09 -0.004107287180320418" free_energy=-28.207973531320143 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000       0.00000000
-Co       0.10563327       1.75738273       1.76055389       3.00000000      -0.00000000       0.00000000      -0.00000000
-Co       1.76055353       0.00000001       1.76055388       3.00000000       0.00000000       0.00000000      -0.00000000
-Co       1.86618681       1.75738272       0.00000001       3.00000000       0.00000000      -0.00000000       0.00000000
