--- examples/data/elasticity_tensor_workflow/defor_8/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_8/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.5211070667208761 -0.0000000019548421 0.0000000009889083 
-lattice_vector 0.2112665468538086 3.5147654359146907 0.0000000222041450 
-lattice_vector 0.0000000023212095 0.0000000221655302 3.5211077603179106 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom 0.1056332745875089 1.7573827290401101 1.7605538912610279 Co
-    initial_moment         3.000000
-atom 1.7605535345210424 0.0000000101053440 1.7605538806534096 Co
-    initial_moment         3.000000
-atom 1.8661868067873417 1.7573827169799243 0.0000000115965267 Co
-    initial_moment         3.000000
