--- examples/data/elasticity_tensor_workflow/defor_9/defor_9.xyz
@@ -1,6 +1,0 @@
-4
-Lattice="3.521107066858156 -1.9583704e-09 9.887105e-10 -1.5141035e-09 3.521109146866583 2.21997037e-08 0.0704221561952355 2.22055362e-08 3.520403468329604" Properties=species:S:1:pos:R:3:initial_magmoms:R:1:forces:R:3 energy=-28.25393979178874 stress="-0.0007653489320759745 -4.34321834236509e-10 0.013051132504771332 -4.34321834236509e-10 -0.00044309728403244274 8.219036098703516e-09 0.013051132504771332 8.219036098703516e-09 -0.00128739277812921" free_energy=-28.25393979178874 pbc="T T T"
-Co      -0.00000000      -0.00000000       0.00000000       3.00000000      -0.00000000      -0.00000000      -0.00000000
-Co       0.03521108       1.76055458       1.76020175       3.00000000      -0.00000000       0.00000000       0.00000000
-Co       1.79576461       0.00000001       1.76020173       3.00000000       0.00000000      -0.00000000       0.00000000
-Co       1.76055353       1.76055457       0.00000001       3.00000000       0.00000000       0.00000000      -0.00000000
