--- examples/data/elasticity_tensor_workflow/defor_9/geometry.in
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-#===============================================================================
-# Created using the Atomic Simulation Environment (ASE)
-
-# Fri Mar  6 15:37:07 2026
-
-=======
-#=======================================================
-# FHI-aims file: C:\Users\akash\OneDrive - Cardiff University\Desktop\elasticity_tensor_calculations\test_mace_calc\defor_9/geometry.in
-# Created using the Atomic Simulation Environment (ASE)
-# Wed Dec  3 17:07:38 2025
->>>>>>> origin/elasticity
-#=======================================================
-lattice_vector 3.5211070668581561 -0.0000000019583704 0.0000000009887105 
-lattice_vector -0.0000000015141035 3.5211091468665829 0.0000000221997037 
-lattice_vector 0.0704221561952355 0.0000000222055362 3.5204034683296039 
-atom -0.0000000000000000 -0.0000000000000000 0.0000000000000001 Co
-    initial_moment         3.000000
-atom 0.0352110773405660 1.7605545845360591 1.7602017452646539 Co
-    initial_moment         3.000000
-atom 1.7957646115266954 0.0000000101235829 1.7602017346591572 Co
-    initial_moment         3.000000
-atom 1.7605535326720256 1.7605545724541063 0.0000000115942071 Co
-    initial_moment         3.000000
